test: rename file to prevent test conflict
