+++ src/electron/background.ts
@@ -5,7 +5,7 @@ import { main } from "@/main";
 import "@/core-setup";
 import "@/ipc/listeners";
 import "@/ipc/story";
-import "@/ipc/test";
+import "@/ipc/testing";
 
 main().then(() => {
 	console.log("started");
